requirements.txt
# Consumer Security Analysis V3 - Python Dependencies

# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
supabase>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Web Scraping
# google-play-scraper>=3.1.0
google-play-scraper>=1.2.7
# app-store-web-scraper>=0.3.0
app-store-web-scraper>=0.2.0
cloudscraper>=1.2.0
selenium>=4.15.0
playwright>=1.40.0

# AI and NLP
openai>=1.0.0
textblob>=0.17.0
spacy>=3.7.0
scikit-learn>=1.3.0
nltk>=3.8.0

# Data Analysis and Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0

# Jupyter and Analysis
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Async and Performance
aiohttp>=3.9.0
asyncio
# concurrent.futures

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0
pytz>=2023.3
tzdata>=2023.3

# Development and Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Optional: For enhanced features
redis>=5.0.0  # For caching
celery>=5.3.0  # For background tasks
fastapi>=0.104.0  # For API endpoints
uvicorn>=0.24.0  # ASGI server
